Created a custom data grid component, created new AdoptionInquiries page with the data grid, and edited the ShelterOverview.jsx so that it uses the custom data grid component
