+++ balto-ui/src/components/ShelterOverview/ShelterOverview.jsx
@@ -1,50 +1,13 @@
 import React from 'react'
 import { useDogRecordsContext } from '../../contexts/dog-records'
 import { useNavigate } from 'react-router-dom'
-import 'bootstrap/dist/css/bootstrap.min.css'
-import { DataGrid, GridToolbarContainer, GridToolbarColumnsButton, GridToolbarFilterButton, GridToolbarExport, useGridApiContext, useGridSelector, gridPageSelector, gridPageCountSelector } from '@mui/x-data-grid'
-import { styled } from '@mui/material/styles'
-import Pagination from '@mui/material/Pagination'
-import PaginationItem from '@mui/material/PaginationItem'
 import Menu from '@mui/material/Menu'
 import MenuItem from '@mui/material/MenuItem'
 import Button from 'react-bootstrap/Button'
-import Form from 'react-bootstrap/Form'
-import { FiMoreHorizontal } from 'react-icons/fi'
+import { Searchbar } from '../CustomDataGrid/CustomDataGrid'
+import CustomDataGrid from '../CustomDataGrid/CustomDataGrid'
+import 'bootstrap/dist/css/bootstrap.min.css'
 import './ShelterOverview.css'
-
-// used to add styling
-const StyledDataGrid = styled(DataGrid)(({ theme }) => ({
-    border: 0,
-    color:
-        theme.palette.mode === 'light' ? 'rgba(0,0,0,.85)' : 'rgba(255,255,255,0.85)',
-    fontFamily: [
-        '-apple-system',
-        'BlinkMacSystemFont',
-        '"Segoe UI"',
-        'Roboto',
-        '"Helvetica Neue"',
-        'Arial',
-        'sans-serif',
-        '"Apple Color Emoji"',
-        '"Segoe UI Emoji"',
-        '"Segoe UI Symbol"',
-    ].join(','),
-    WebkitFontSmoothing: 'auto',
-    letterSpacing: 'normal',
-    '& .MuiDataGrid-columnsContainer': {
-        backgroundColor: theme.palette.mode === 'light' ? '#fafafa' : '#1d1d1d',
-    },
-    '& .MuiDataGrid-columnsContainer, .MuiDataGrid-cell': {
-        borderBottom: `1px solid ${theme.palette.mode === 'light' ? '#f0f0f0' : '#303030'}`,
-    },
-    '& .MuiDataGrid-cell': {
-        color: theme.palette.mode === 'light' ? 'rgba(0,0,0,.85)' : 'rgba(255,255,255,0.65)',
-    },
-    '& .MuiPaginationItem-root': {
-        borderRadius: 5,
-    },
-    }))
 
 export default function ShelterOverview() {
 
@@ -82,15 +45,15 @@ export default function ShelterOverview() {
     return (
         <div className="shelter-overview primary-container" style={{ display: 'flex', height: '100%', width: '100%' }}>
             <div className="top-bar">
-                <DogRecordSearch setSearchQuery={setSearchQuery} />
+                <Searchbar setSearchQuery={setSearchQuery} />
                 <Button href="/admin-dashboard/add-record" className="dog-record-add-btn mb-3">+ New</Button>
             </div>
-            {/* <ShelterOverviewGrid rows={rows} columns={columns} /> */}
             <RowContextMenu rows={rows} columns={columns} />
         </div>
     )
 }
 
+// provides both the context menu and the actual data grid
 export function RowContextMenu(props) {
     const [selectedRow, setSelectedRow] = React.useState()
     const navigate = useNavigate()
@@ -113,7 +76,7 @@ export function RowContextMenu(props) {
   
     return (
       <div style={{ width: '100%' }}>
-        <ShelterOverviewGrid
+        <CustomDataGrid
             rows={props.rows}
             columns={props.columns}
             componentsProps={{
@@ -151,63 +114,4 @@ export function RowContextMenu(props) {
       </div>
     )
   }
-
-export function ShelterOverviewGrid(props) {
-return (
-    <div style={{ flexGrow: 1}} className="shelter-overview-grid">
-        <StyledDataGrid
-            rows={props.rows}
-            columns={props.columns}
-            rowHeight={70}
-            autoHeight
-            pagination
-            rowsPerPageOptions={[30]}
-            components={{
-                Toolbar: CustomToolbar,
-                Pagination: CustomPagination
-            }}
-            componentsProps={props.componentsProps}
-        />
-    </div>
-    )
-}
-
-// searchbar updates searchQuery var whenever the value is changed
-export function DogRecordSearch({ setSearchQuery }) {
-    return (
-        <Form.Group className="mb-3 dog-record-search">
-          <Form.Control placeholder="Search by Name" onChange={(evt)=>setSearchQuery(evt.target.value)} />
-        </Form.Group>
-    )
-}
-
-// toolbar with column list, filters, and export option
-export function CustomToolbar() {
-    return (
-      <GridToolbarContainer>
-        <GridToolbarColumnsButton />
-        <GridToolbarFilterButton />
-        <GridToolbarExport />
-      </GridToolbarContainer>
-    )
-  }
-
-// custom pagination that appears at the footer
-export function CustomPagination() {
-    const apiRef = useGridApiContext()
-    const page = useGridSelector(apiRef, gridPageSelector)
-    const pageCount = useGridSelector(apiRef, gridPageCountSelector)
-
-    return (
-        <Pagination
-          color="primary"
-          variant="outlined"
-          count={pageCount}
-          page={page + 1}
-          shape="rounded"
-          renderItem={(props2) => <PaginationItem {...props2} disableRipple />}
-          onChange={(event, value) => apiRef.current.setPage(value - 1)}
-        />
-      );
-}
